backend/requirements.txt
pandas==2.2.3
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic[email]==2.6.4
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
pymongo==4.6.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose==3.3.0